Add Gwmi-Async.ps1
A lot of improvements can be made, but it's still tremendously useful on occasion. To do's are stuff like CIM support and avoiding all the string concatenation, and other performance improvements. Especially creating the XML is time-consuming with the code as-is. Haven't had opportunity to review yet (or sufficient need).